requirements.txt
Flask==0.10.1
PyYAML==3.11
requests==2.3.0
